refactor and run server first when testing
