+++ src/startServer.ts
@@ -1,0 +1,18 @@
+import { importSchema } from "graphql-import";
+import { GraphQLServer } from "graphql-yoga";
+import { resolvers } from "./resolvers";
+import * as path from "path";
+import { createTypeormConnection } from "./utils/createTypeormConnection";
+// ... or using `require()`
+// const { GraphQLServer } = require('graphql-yoga')
+
+export const startServer = async () => {
+  const typeDefs = importSchema(path.join(__dirname, "schema.graphql"));
+  const server = new GraphQLServer({ typeDefs, resolvers });
+  await createTypeormConnection();
+  const app = await server.start({
+    port: process.env.NODE_ENV === "test" ? 0 : 4000
+  });
+  console.log("Server is running on localhost:4000");
+  return app;
+};
