+++ src/tests/register.test.ts
@@ -1,11 +1,14 @@
 import { request } from "graphql-request";
-import { host } from "./constant";
 import { User } from "../entity/User";
-import { createTypeormConnection } from "../utils/createTypeormConnection";
+import { startServer } from "../startServer";
 
-beforeAll(async ()=>{
-  await createTypeormConnection();
-})
+let getHost = () => "";
+
+beforeAll(async () => {
+  const app = await startServer();
+  const { port } = app.address(); 
+  getHost = () => `http://127.0.0.1:${port}`
+});
 
 const email = "zei@han.com";
 const password = "123";
@@ -17,12 +20,11 @@ mutation {
 `;
 
 test("Register user", async () => {
-  const response = await request(host, mutation);
+  const response = await request(getHost(), mutation);
   expect(response).toEqual({ register: true });
-  const users = await User.find({where:{email}});
+  const users = await User.find({ where: { email } });
   expect(users).toHaveLength(1);
   const user = users[0];
   expect(user.email).toEqual(email);
   expect(user.password).not.toEqual(password);
-
 });
